requirements.txt
pip==25.3
pytesseract==0.3.13
pytest==9.0.0
selenium-wire==5.1.0
setuptools==80.9.0
webdriver-manager==4.0.0
wheel==0.45.1
allure-pytest==2.15.0
yagmail==0.15.293
pytest-json-report==1.5.0
pyminizip==0.2.6
